api/requirements.txt
tensorflow==2.16.1
keras==3.1.1
keras-cv==0.8.2
fastapi==0.110.0
uvicorn==0.29.0
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
boto3==1.34.74